# swagger.yaml
swagger: "2.0"
info:
  description: "API for querying parking rates"
  version: "1.0.0"
  title: "ParkingChampion"
tags:
- name: "rates"
  description: "Get and set the current parking rate structure"
- name: "quotes"
  description: "Get a parking rate for a specific start and end time"
schemes:
- "http"
paths:
  /rates:
    post:
      tags:
      - "rates"
      summary: "Set the current parking rates"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New rates"
        required: true
        schema:
          $ref: "#/definitions/Rates"
      responses:
        400:
          description: "One of the provided rates was invalid"
    get:
      tags:
        - "rates"
      summary: "Get the current parking rates"
      responses:
        200:
          description: OK
          examples:
            application/json: {
              "rates": [
                {
                  "price": "1000",
                  "days": "mon,tues,thurs",
                  "times": "0900-2100",
                  "tz": "America/Chicago"
                }
              ]
            }
  /getRate:
    get:
      tags:
        - "quotes"
      summary: "Get the applicable parking rate"
      parameters:
        - in: query
          name: start
          type: string
          required: true
          description: Timestamp that specifies the desired beginning parking time (ex. 2015-07-04T15:00:00+00:00)
        - in: query
          name: end
          type: string
          required: true
          description: Timestamp that specifies the desired end parking time
            (ex. 2015-07-04T20:00:00+00:00)
      responses:
        200:
          description: OK
          examples:
            application/json: {"rate":"1750"}
        400:
          description: "Start/End times either were not provided or were invalid"
definitions:
  Rates:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      rates:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Rate"
  Rate:
    type: "object"
    properties:
      price:
        type: "integer"
        example: "1000"
      days:
        type: "string"
        example: "mon,tues,thurs"
      times:
        type: "string"
        example: "0900-2100"
      tz:
        type: "string"
        example: "America/Chicago"